Check Python snippets too, and rename the script
Python is parsed with ast.parse rather than run, so a fragment is fine as
long as it stands on its own; only blocks that open mid-indentation are
skipped, since those were lifted out of a function. A syntax error is
reported as file:line:col with the offending line, mapped back to the
.mdx, instead of a parser traceback.

check-cpp-snippets becomes check-code-snippets, with the workflow and
yarn script renamed to match.

Run over the General section — 29 module files and all 8 of its internal
solutions — all 59 C++ snippets and all 29 Python ones check out.
