+++ package.json
@@ -13,7 +13,7 @@
     "lint": "next lint",
     "clean": "rimraf dist .next data public/usaco-divisions.json public/problem-tags.json",
     "check-ts-errors": "tsc --noEmit",
-    "check-cpp-snippets": "yarn tsx scripts/check-cpp-snippets.ts",
+    "check-code-snippets": "yarn tsx scripts/check-code-snippets.ts",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
     "chromatic": "chromatic --project-token 6f3c58ad349c --exit-zero-on-changes",
